Move \Nova\Events\PriorityQueue into namespace \Nova\Utils
